External structures
External structures are looked up using the include paths.
For that the code model was extended to call a lookup function if no type was found.
The include paths are retrieved either by argument or using CMake. Therefore CMake is now configured when executing the generate command.
The include paths are cached beside the settings of the CLI. Every time a code model is build they are updated. Every missing or outdated file is removed from the cache.
For performance reasons the caching of file contents for the UnDelete functionality is delayed until the content is changed or the file gets deleted.
#exported-commit:5fe19cd69daac3f52fd2c1584edc9dca9fafce56